+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+### Changed
+
+- Some better support for indenting mixed content when whitespace is set to ignore.
+
 ## [0.9.0] - 2020-07-21
 
 ### Added